refactor: move functionality to `render.tsx`
chore: removed `tw.ts` and let user handle twind install
