+++ x/x.test.tsx
@@ -1,18 +1,9 @@
 // Copyright 2023 Samuel Kopp. All rights reserved. Apache-2.0 license.
 /** @jsx h */
 import { cheetah } from '../cheetah.ts'
-import {
-  assert,
-  assertEquals,
-  assertInstanceOf,
-  defineConfig,
-  DOMParser,
-  presetAutoPrefix,
-  presetTailwind,
-} from '../test/deps.ts'
+import { assertEquals, assertInstanceOf } from '../test/deps.ts'
 import { h, jsx } from './jsx.tsx'
 import { createKey, importKey, sign, verify } from './jwt.ts'
-import install from './tw.ts'
 
 // TODO delete at v2.0
 
@@ -57,44 +48,6 @@ Deno.test('x', async (t) => {
       )
       assertEquals(
         b.headers.get('content-type'),
-        'text/html; charset=utf-8',
-      )
-    })
-
-    await t.step('styled', async () => {
-      const app = new cheetah()
-
-      const twindConfig = defineConfig({
-        presets: [presetAutoPrefix(), presetTailwind()],
-      })
-
-      install(twindConfig)
-
-      function Styled() {
-        return (
-          <h3 class='text-sm italic' id='styled'>
-            styled <code class='font-mono'>h3</code> component
-          </h3>
-        )
-      }
-
-      app.get('/a', (c) => jsx(c, Styled))
-
-      const a = await app.fetch(new Request('http://localhost/a'))
-
-      const document = new DOMParser().parseFromString(
-        await a.text(),
-        'text/html',
-      )
-
-      assert(document)
-      assert([...document.getElementsByTagName('style')].length)
-      assertEquals(
-        document.getElementById('styled')?.innerText,
-        'styled h3 component',
-      )
-      assertEquals(
-        a.headers.get('content-type'),
         'text/html; charset=utf-8',
       )
     })
